requirements.txt
discord.py
python-dotenv
imap-tools
